Seperated files into a main file and classes file for better organization
