+++ main.py
@@ -57,78 +57,26 @@ This is intended to be the logic backend for War.
 '''
 
 import random
+import warClasses
+
+print('Welcome to WAR!')
+player1 = input('Please enter a name for player 1: ')
+print('Hello ' + player1)
+player2 = input('Please enter a name for player 2: ')
+print('Hello ' + player2)
 
 
-
-class Card():
-	def __init__(self, suit, val):
-		self.suit = suit
-		self.val  = val
-
-	def show(self):
-		print("{} of {}".format(self.val, self.suit))
-
-#Builds a deck of cards
-class Deck():
-	def __init__(self):
-		self.cards = []
-		self.build()
-
-	def build(self):
-		for cardsuit in ["Diamonds", "Spades", "Hearts", "Clubs"]:
-			for cardvalue in range(1, 14): 
-				self.cards.append(Card(cardsuit, cardvalue))
-
-	def show(self):
-		for c in self.cards:
-			c.show()
-
-    #Shuffles the deck
-	def shuffle(self):
-		for i in range(len(self.cards)-1, 0, -1):
-			r = random.randint(0 , i)
-			self.cards[i], self.cards[r] = self.cards[r], self.cards[i]	
-
-	#Splits the deck into even halves for 2 players
-	def deal(self):
-		for i in self.cards:
-			self.firstHalf = self.cards[:26]
-			self.secondHalf = self.cards[26:]
-
-	#Verifies that the deck has been dealt correctly.
-	def showHands(self):
-		for i in self.firstHalf:
-		  i.show()
-		for i in self.secondHalf:
-		  i.show()
-			
-
-class Player():
-	def __init__(self, name):
-		self.name = name
-		self.hand = []
- 
-
-	def playerHand(self):
-	  x = 0
-	  while(x < 26):
-		  self.hand.append(deck.cards.pop())
-		  x += 1
-		  
-	def show(self):
-	  for i in self.hand:
-	    i.show()
 
 #testcard = Card('Clubs', 6)
 #testcard.show()
-deck = Deck()
-deck.shuffle()
-deck.deal()
+#deck = Deck()
+#deck.shuffle()
+#deck.deal()
 #deck.show()
-Bob = Player('Bob')
-Bob.playerHand()
-Bob.show()
-print('\n')
-Carl = Player('Carl')
-Carl.playerHand()
-Carl.show()
+#Bob = Player('Bob')
+#Bob.playerHand()
+#Bob.show()
+#print('\n')
+#Carl = Player('Carl')
+#Carl.playerHand()
+#Carl.show()
